corretto esercizio 5-2

+++ exercises/ex5-2/main.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 
 int main() {
-	CMatrix<double> A, B;
+	CMatrix<int> A, B;
 	A.read_from_file("A.txt");
 	B.read_from_file("B.txt");
 	auto C = A * B;
